Python-Service/requirements.txt
fastapi
uvicorn
httpx
apscheduler
prometheus-client
pytest
pydantic-settings